(한우석) edit : socket 헤더 컴포넌트에서 연결 하도록 수정
