+++ src/pages/Mypage.js
@@ -3,33 +3,16 @@ import styled from 'styled-components';
 
 import { useDispatch, useSelector } from 'react-redux';
 import { loadMyCardDB } from '../features/cards/actions';
-// import { history } from '../features/configureStore';
 
 import CardProfile from '../components/CardProfile';
 import CardAdd from '../components/CardAdd';
 import CardBackWrite from '../components/CardBackWrite';
-import useSocketNotiRoom from '../hooks/useSocketNotiRoom';
 
 import { Grid, Text } from '../elements';
 
 const Mypage = () => {
   const dispatch = useDispatch();
-  const [wsConnectSubscribe, wsDisConnectUnsubscribe, token] =
-    useSocketNotiRoom();
-
-  React.useEffect(() => {
-    if (!token) {
-      return null;
-    }
-    wsConnectSubscribe();
-
-    return () => {
-      wsDisConnectUnsubscribe();
-    };
-  }, []);
-
   const cardCount = useSelector((state) => state.cards.allIds);
-  console.log(cardCount.length);
 
   const [click, setClick] = React.useState(false);
